+++ js/back-handler.js
@@ -1,0 +1,9 @@
+import controller from './controller';
+
+export default (screen, state) => {
+  const backButton = screen.querySelector(`.back`);
+
+  backButton.addEventListener(`click`, () => {
+    controller(state, {back: true});
+  });
+};
