--- js/return-behavior.js
@@ -1,10 +1,0 @@
-import changeScreen from './render-screen';
-import renderBack from './screen-greeting';
-
-export default (screen) => {
-  const backButton = screen.querySelector(`.back`);
-
-  backButton.addEventListener(`click`, () => {
-    changeScreen(renderBack());
-  });
-};
